Replace nosetests w py.test; mv test to package
Switched to py.test because nose is not developed anymore. Their own
site recommends using py.test.

The tests are now part of the fatiando package and are installed with it
in the `fatiando.tests` package. You can use

    python -c "import fatiando; fatiando.test()"

to run the tests after installing. This will still require py.test.

Moved extra requirements files to `ci` folder and use only those two to
install in TravisCI.

TODO:

* document the change: how to test in devel; how to test after install